+++ Arduino/changelog.md
@@ -1,11 +1,15 @@
 # Changelog Arduino
 
-### v1.0.2 - 26-09-2020
+### v1.0.2 - 27-09-2020
 Fixes:
 <ul>
 <li>Included all required libraries in project folder</li>
 <li>Cleaned up the code</li>
 <li>Simplified and improved sensitivity settings</li>
+</ul>
+Additions:
+<ul>
+<li>Added X and Y compensation to fine-tune the measured coordinate</li>
 </ul>
 
 ### v1.0.1 - 24-09-2020