+++ Module/changelog.md
@@ -1,3 +1,18 @@
 # Changelog Module
+### v1.0.1 - 27-09-2020
+Fixes:
+<ul>
+<li>Simplified and improved sensitivity settings</li>
+<li>Fixed issue where the calibration menu would not update properly</li>
+<li>Bigger tokens now center properly</li>
+<li>Multipoint calibration fixed, multipoint offset added</li>
+</ul>
+Additions:
+<ul>
+<li>Added X and Y compensation to fine-tune the measured coordinate</li>
+<li>Added calibration and offset checkboxes to calibration menu</li>
+<li>Added low battery notification</li>
+</ul>
+
 ### v1.0.0 - 22-09-2020
 Initial Release